+++ mpdm_d.c
@@ -34,12 +34,12 @@
 
 /** data **/
 
-static wchar_t *dump_1(const mpdm_t v, int l, wchar_t *ptr, size_t *size);
-wchar_t *(*mpdm_dump_1) (const mpdm_t v, int l, wchar_t *ptr, size_t *size) = NULL;
+static wchar_t *dump_1(const mpdm_t v, int l, wchar_t *ptr, int *size);
+wchar_t *(*mpdm_dump_1) (const mpdm_t v, int l, wchar_t *ptr, int *size) = NULL;
 
 /** code **/
 
-static wchar_t *dump_1(const mpdm_t v, int l, wchar_t *ptr, size_t *size)
+static wchar_t *dump_1(const mpdm_t v, int l, wchar_t *ptr, int *size)
 /* dumps one value to the ptr dynamic string with 'l' indenting level */
 {
     int n;
@@ -53,7 +53,7 @@ static wchar_t *dump_1(const mpdm_t v, int l, wchar_t *ptr, size_t *size)
 
     if (v != NULL) {
         char tmp[256];
-        size_t s;
+        int s;
         int c = 0;
         mpdm_t w, i;
         wchar_t *str;
@@ -97,7 +97,7 @@ static wchar_t *dump_1(const mpdm_t v, int l, wchar_t *ptr, size_t *size)
 }
 
 
-static wchar_t *do_dump(mpdm_t v, size_t *size)
+static wchar_t *do_dump(mpdm_t v, int *size)
 {
     wchar_t *ptr;
 
@@ -121,7 +121,7 @@ static wchar_t *do_dump(mpdm_t v, size_t *size)
  */
 mpdm_t mpdm_dumper(const mpdm_t v)
 {
-    size_t size = 0;
+    int size = 0;
     wchar_t *ptr;
 
     ptr = do_dump(v, &size);
@@ -139,7 +139,7 @@ mpdm_t mpdm_dumper(const mpdm_t v)
  */
 void mpdm_dump(const mpdm_t v)
 {
-    size_t size = 0;
+    int size = 0;
     wchar_t *ptr;
 
     ptr = do_dump(v, &size);
